backend/requirements.txt
Flask==2.3.3
flask_cors==5.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.55.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
MarkupSafe==2.1.3